+++ apps/app-frontend/src/components/multiplayer/servers/ServersOverview.vue
@@ -1,8 +1,7 @@
 <script setup lang="ts">
-import { setEulaAccepted } from '@modrinth/server'
 import { PlusIcon, RefreshCwIcon, ServerIcon, SpinnerIcon } from '@modrinth/assets'
 import { ButtonStyled, defineMessages, EmptyState, useVIntl } from '@modrinth/ui'
-import { onMounted, ref, useTemplateRef } from 'vue'
+import { onMounted, useTemplateRef } from 'vue'
 import type { ComponentExposed } from 'vue-component-type-helpers'
 import { useRouter } from 'vue-router'
 
@@ -10,15 +9,13 @@ import CreateServerModal from '@/components/multiplayer/servers/CreateServerModa
 import EulaModal from '@/components/multiplayer/servers/EulaModal.vue'
 import ServerCard from '@/components/multiplayer/servers/ServerCard.vue'
 import { type ServerView, useServers } from '@/composables/useServers'
-import { servers as serversApi } from '@/helpers/servers'
+import { useServerLifecycle } from '@/composables/useServerLifecycle'
 
 const router = useRouter()
 const { formatMessage } = useVIntl()
-const { servers, isRefreshing, refresh, startServer, stopServer } = useServers()
+const { servers, isRefreshing, refresh, stopServer } = useServers()
+const { eulaModal, eulaText, tryStartServer, acceptEula, declineEula } = useServerLifecycle()
 const createModal = useTemplateRef<ComponentExposed<typeof CreateServerModal>>('createModal')
-const eulaModal = useTemplateRef<ComponentExposed<typeof EulaModal>>('eulaModal')
-const eulaText = ref('')
-const eulaPendingId = ref('')
 
 const messages = defineMessages({
 	create: { id: 'app.servers.create.title', defaultMessage: 'Create server' },
@@ -52,38 +49,6 @@ async function toggleRunning(server: ServerView) {
 	} else {
 		await tryStartServer(server)
 	}
-}
-
-/** Starts the server; if the EULA is unaccepted, shows the EULA modal first. */
-async function tryStartServer(server: ServerView) {
-	if (!server.eulaAccepted && server.eulaExists) {
-		try {
-			eulaText.value = await serversApi.readFile(server.id, 'eula.txt')
-			eulaPendingId.value = server.id
-			eulaModal.value?.show()
-			return
-		} catch {
-			// No eula.txt: a fresh start will generate it
-		}
-	}
-	await startServer(server.id)
-}
-
-async function onEulaAccept() {
-	const id = eulaPendingId.value
-	if (!id) return
-	try {
-		const updated = setEulaAccepted(eulaText.value, true)
-		await serversApi.writeFile(id, 'eula.txt', updated)
-		eulaModal.value?.hide()
-		await startServer(id)
-	} catch (error) {
-		console.error(error)
-	}
-}
-
-function onEulaDecline() {
-	eulaModal.value?.hide()
 }
 </script>
 
@@ -144,6 +109,6 @@ function onEulaDecline() {
 		</div>
 
 		<CreateServerModal ref="createModal" />
-		<EulaModal ref="eulaModal" :text="eulaText" @accept="onEulaAccept" @decline="onEulaDecline" />
+		<EulaModal ref="eulaModal" :text="eulaText" @accept="acceptEula" @decline="declineEula" />
 	</div>
 </template>
